% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adverse_events.R
\name{adverse_events}
\alias{adverse_events}
\title{Create a dataframe of adverse events from a tumor registry that can be combined into a storyboard}
\usage{
adverse_events(data)
}
\arguments{
\item{data}{is a data frame which contains the data for which you want to create a storyboard}
}
\value{
A data frame with five variables ("record_id", "description", "value", "date", and "hover" that can be combined with others dfs with the same five variables to form a storyboard
}
\description{
`adverse_events()`wrangles data from the Adverse Events form of tumor registries to produce a dataframe of details about adverse events of systemic therapy, which can then be incorporated into a Patient Storyboard
}
\examples{
# Test with embedded data set "storyboard_dataset"
storyboard_dataset \%>\%
  adverse_events()
}
