% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ae_codebook.R
\name{ae_codebook}
\alias{ae_codebook}
\title{Replace the data in the Adverse Events Instrument with the strings from the codebook}
\usage{
ae_codebook(data)
}
\arguments{
\item{data}{is a dataframe from a tumor registry with the Adverse Events Instrument}
}
\value{
A dataframe which contains the data with numerics replaced by stings
}
\description{
`ae_codebook()`transforms adverse event data to the appropriate strings per the AE codebook
}
\examples{
# Test with embedded data set "storyboard_dataset"
storyboard_dataset \%>\%
  ae_codebook()
}
