% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_storyboard_dfs.R
\name{combine_storyboard_dfs}
\alias{combine_storyboard_dfs}
\title{Combine Storyboard dataframes}
\usage{
combine_storyboard_dfs(data)
}
\arguments{
\item{data}{is a data frame which contains the data for which you want to create a storyboard}
}
\value{
A data frame that has combined the various dataframes of the storyboard package
}
\description{
`combine_storyboard_dfs()` combine the various Storybards DFs
}
\examples{
# Test with embedded data set "storyboard_dataset"
storyboard_dataset \%>\%
  combine_storyboard_dfs()
}
