% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date.shift.df.R
\name{date.shift.df}
\alias{date.shift.df}
\title{Randomly date shift the combined storyboard data frame}
\usage{
date.shift.df(data)
}
\arguments{
\item{data}{a data frame that is downstream of combine_storyboards_df()}

\item{date}{the name of the column with the date intended to be shifted. Defaults to "date".}
}
\value{
A data frame of shifted dates
}
\description{
`date.shift.df()` shifts the dates a unified random number of weeks either forward or back between 1 and 52
}
\examples{
# Test with embedded data set "storyboard_dataset"
storyboard_dataset \%>\%
  date.shift.df()

}
