% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateShift.R
\name{dateShift}
\alias{dateShift}
\title{Shift dates a number of days, bizdays, months, weeks or years. This code was taking directly from TimeWarp, since it is no longer on CRAN}
\usage{
dateShift(
  x,
  by = "days",
  k.by = 1,
  direction = 1,
  holidays = NULL,
  silent = FALSE,
  optimize.dups = TRUE
)
}
\description{
`dateShift()` shifts dates a number of days, bizdays, months, weeks or years.
}
