% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Demographics Data Frame.R
\name{demographics_df}
\alias{demographics_df}
\title{Create a Data Frame of Patient Demographics}
\usage{
demographics_df(data, record_id)
}
\arguments{
\item{data}{The data frame from the MCC Patient registry. Required.}

\item{record_id}{The record id of the subject of interest. Required.}
}
\value{
a data frame
}
\description{
`demographics_df()` creates a table of patient demographics
}
