% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_record.R
\name{filter_by_record}
\alias{filter_by_record}
\title{Filter the storyboards dataframe by the subject of interest}
\usage{
filter_by_record(data, id)
}
\arguments{
\item{data}{is a data frame of the combined storyboard (e.g. after combine_storyboard_dfs() is called)}

\item{id}{is a character vector corresponding to a subject id: e.g. "7612-1"}
}
\value{
The storyboard data frame filtered by the record of interest
}
\description{
`filter_by_record()` filters the storyboards dataframe by the subject of interest
}
\examples{
# Test with embedded data set "storyboard_dataset"
storyboard_dataset \%>\%
  filter.by.record()

}
