% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomics.R
\name{genomics}
\alias{genomics}
\title{Create a dataframe of genomic data from a tumor registry that can be combined into a storyboard}
\usage{
genomics(data)
}
\arguments{
\item{data}{is a data frame which contains the data for which you want to create a storyboard}
}
\value{
A data frame with five variables ("record_id", "description", "value", "date", and "hover" that can be combined with others dfs with the same five variables to form a storyboard
}
\description{
`genomics()` wrangles data from the Genoimcs form of tumor registries to produce a dataframe of details about genomic data from tumors or blood, which can then be incorporated into a Patient Storyboard
}
\examples{
# Test with embedded data set "storyboard_dataset"
storyboard_dataset \%>\%
  genomics()

}
