% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Genomics Data Frame.R
\name{genomics_df}
\alias{genomics_df}
\title{Create a Data Frame of Genomics}
\usage{
genomics_df(data)
}
\arguments{
\item{data}{The data frame from the MCC Patient registry. Required.}
}
\value{
a data frame
}
\description{
`genomics_df()` creates a table of genomic information
}
