% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_StoryboardR.R
\name{launch_StoryboardR}
\alias{launch_StoryboardR}
\title{Lauch StoryboardR App}
\usage{
launch_StoryboardR(Data, DateShift = FALSE, ...)
}
\arguments{
\item{Data}{a data frame contains raw data from an exported REDCap project.}

\item{DateShift}{If TRUE, dates will be shifted a unified random number of weeks either forward or back between 1 and 52. Defaults to FALSE.}
}
\description{
launch_StoryboardR()` launches the StoryboardR shiny application.
}
