% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lesions Data Frame.R
\name{lesions_df}
\alias{lesions_df}
\title{Create a Data Frame of MCC Lesions in Patients}
\usage{
lesions_df(data)
}
\arguments{
\item{data}{The data frame from the MCC Patient registry. Required.}
}
\value{
a data frame
}
\description{
`lesions_df()` creates a table of MCC lesions experienced by the patient
}
