% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathological_staging.R
\name{pathological_staging}
\alias{pathological_staging}
\title{Create a dataframe of pathological staging from a tumor registry that can be combined into a storyboard}
\usage{
pathological_staging(data)
}
\arguments{
\item{data}{is a data frame which contains the data for which you want to create a storyboard}
}
\value{
A data frame with five variables ("record_id", "description", "value", "date", and "hover" that can be combined with others dfs with the same five variables to form a storyboard
}
\description{
`pathological_staging()`wrangles data from the Presentation and Initial Staging form of tumor registries to produce a dataframe of details about the initial pathological staging, which can then be incorporated into a Patient Storyboard
}
\examples{
# Test with embedded data set "storyboard_dataset"
storyboard_dataset \%>\%
  pathological_staging()

}
