% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Staging Data Frame.R
\name{staging_df}
\alias{staging_df}
\title{Create a Data Frame of Initial Staging for MCC patients}
\usage{
staging_df(data)
}
\arguments{
\item{data}{The data frame from the MCC Patient registry. Required.}
}
\value{
a data frame with four columns: clinical and pathological staging at presentation and the dates staged
}
\description{
`staging_df()` creates a table of initial staging at presentation
}
