% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storyboard_Plot.R
\name{storyboard_plot}
\alias{storyboard_plot}
\title{Create a plotly interactive timeline of a patient journey from a tumor registry}
\usage{
storyboard_plot(data)
}
\arguments{
\item{data}{is a data frame down stream of combine_storyboard_df and further filtered for a specific record ID of interest}
}
\value{
A ggplotly Storyboard for a given subject
}
\description{
`storyboard_plot()` takes the aggregated data frames from combine_storyboard_dfs to produce a plotly data visualization of a patient journey
}
\examples{
# Test with embedded data set "storyboard_dataset"
StoryboardR::storyboard_dataset \%>\%

}
