% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subject Status Data Frame.R
\name{subject_status_df}
\alias{subject_status_df}
\title{Create a Data Frame of Subject Status}
\usage{
subject_status_df(data)
}
\arguments{
\item{data}{The data frame from the MCC Patient registry. Required.}
}
\value{
a data frame
}
\description{
`subject_status_df()` creates a table of subject status information
}
