% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Surgery Data Frame.R
\name{surgery_df}
\alias{surgery_df}
\title{Create a Data Frame of Surgical Therapy}
\usage{
surgery_df(data)
}
\arguments{
\item{data}{The data frame from the MCC Patient registry. Required.}
}
\value{
a data frame
}
\description{
`surgery_df()` creates a table of Surgical Therapy information
}
