% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Systemic Therapy Data Frame.R
\name{systemic_therapy_df}
\alias{systemic_therapy_df}
\title{Create a Data Frame of Systemic Therapy}
\usage{
systemic_therapy_df(data)
}
\arguments{
\item{data}{The data frame from the MCC Patient registry. Required.}
}
\value{
a data frame
}
\description{
`systemic_therapy_df()` creates a table of Systemic Therapy information
}
