% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XRT Data Frame.R
\name{xrt_df}
\alias{xrt_df}
\title{Create a Data Frame of Radiation Therapy}
\usage{
xrt_df(data)
}
\arguments{
\item{data}{The data frame from the MCC Patient registry. Required.}
}
\value{
a data frame
}
\description{
`xrt_df()` creates a table of Radiation Therapy information
}
