% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Storyboard_Plot.R
\name{storyboard_plot}
\alias{storyboard_plot}
\title{Create a plotly interactive timeline of a patient journey from a tumor registry}
\usage{
storyboard_plot(data)
}
\arguments{
\item{data}{is a data frame which contains the data for which you want to create a storyboard}
}
\value{
A ggplotly Storyboard for a given subject
}
\description{
`storyboard_plot()` provides natural language processing tools to abstract data from a variety of genomic reports and import them to REDCap
}
\examples{
# Test with embedded data set "storyboard_dataset"
storyboard_dataset \%>\%
  storyboard_plot(id = "Simulated Patient 1")

}
